+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.0.45
+
+* `Add support of containers in `UIKitContainer` modifiers and `Container` function
+
 ## 0.0.44
 
 * `Versions`: